Refactor: legacy -> 변경된 패키지 경로로 변경
