+++ src/main/java/com/ssafy/enjoytrip/attraction/controller/AttractionController.java
@@ -1,6 +1,6 @@
 package com.ssafy.enjoytrip.attraction.controller;
 
-import com.ssafy.enjoytrip.legacy.service.AttractionService;
+import com.ssafy.enjoytrip.attraction.service.AttractionService;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
