--- src/main/java/com/ssafy/enjoytrip/legacy/controller/SearchAttractionController.java
@@ -1,27 +1,0 @@
-package com.ssafy.enjoytrip.legacy.controller;
-
-import com.ssafy.enjoytrip.legacy.service.AttractionService;
-import java.util.List;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("api/v1/search")
-public class SearchAttractionController {
-
-    private final AttractionService attractionService;
-
-    public SearchAttractionController(AttractionService attractionService) {
-        this.attractionService = attractionService;
-    }
-
-    @GetMapping("/{sidoCode}")
-    public ResponseEntity<List<String>> getGugun(@PathVariable("sidoCode") int sidoCode) {
-        List<String> gugunList = attractionService.getGugun(sidoCode);
-        return new ResponseEntity<List<String>>(gugunList, HttpStatus.OK);
-    }
-}
