--- src/main/java/com/ssafy/enjoytrip/legacy/controller/TravelReviewBoardController.java
@@ -1,8 +1,0 @@
-package com.ssafy.enjoytrip.legacy.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TravelReviewBoardController {
-
-}
