--- src/main/java/com/ssafy/enjoytrip/legacy/controller/request/LoginRequest.java
@@ -1,19 +1,0 @@
-package com.ssafy.enjoytrip.legacy.controller.request;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Getter
-public class LoginRequest {
-    @NotBlank(message = "이메일은 필수 입력사항입니다.")
-    @Email(message = "이메일 형식을 맞춰주세요.")
-    private String email;
-
-    @NotBlank(message = "비밀번호는 필수 입력사항입니다.")
-    private String password;
-
-
-}
