--- src/main/java/com/ssafy/enjoytrip/legacy/dto/attraction/Gugun.java
@@ -1,17 +1,0 @@
-package com.ssafy.enjoytrip.legacy.dto.attraction;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class Gugun {
-
-    private int code;
-    private String name;
-    private int sidoCode;
-}
