--- src/main/java/com/ssafy/enjoytrip/legacy/dto/trb/Comment.java
@@ -1,22 +1,0 @@
-package com.ssafy.enjoytrip.legacy.dto.trb;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class Comment {
-
-    private int id;
-    private int userId;
-    private int trbId;
-    private String content;
-    private LocalDateTime createtime;
-    private LocalDateTime updatetime;
-}
