--- src/main/java/com/ssafy/enjoytrip/legacy/dto/user/UserRole.java
@@ -1,7 +1,0 @@
-package com.ssafy.enjoytrip.legacy.dto.user;
-
-public enum UserRole {
-    ADMIN,
-    COMMON;
-
-}
