--- src/main/java/com/ssafy/enjoytrip/legacy/exception/UserDuplicatedNicknameException.java
@@ -1,5 +1,0 @@
-package com.ssafy.enjoytrip.legacy.exception;
-
-public class UserDuplicatedNicknameException extends RuntimeException {
-
-}
