--- src/main/java/com/ssafy/enjoytrip/legacy/mapper/SearchAttractionMapper.java
@@ -1,11 +1,0 @@
-package com.ssafy.enjoytrip.legacy.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-import java.util.List;
-
-@Mapper
-public interface SearchAttractionMapper {
-
-    List<String> getGugun(int sidoCode);
-}
