--- src/main/java/com/ssafy/enjoytrip/legacy/service/AttractionService.java
@@ -1,8 +1,0 @@
-package com.ssafy.enjoytrip.legacy.service;
-
-import java.util.List;
-
-public interface AttractionService {
-
-    List<String> getGugun(int sidoCode);
-}
