+++ src/main/java/com/ssafy/enjoytrip/user/dto/request/CreateUserRequest.java
@@ -1,6 +1,6 @@
 package com.ssafy.enjoytrip.user.dto.request;
 
-import com.ssafy.enjoytrip.legacy.dto.user.User;
+import com.ssafy.enjoytrip.user.dto.User;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
